Extracted setFixedRange to StockUtilities.
